adjust handling and rendering of GC events
Recognize the updated `gc-info` mode field, and change the way a GC
even tis rendered so that it uses the start and end time in the event.
Allow the rawn region for a variable-width event (like a GC) to be
narrower than the normal minimum event width to better reflect the
relative duration of events.